# gamemaster-ai/docs/OPENAPI.yaml
openapi: 3.0.3
info:
  title: GameMaster AI API
  version: 0.2.0
  description: Auto-synced API surface generated from app/api route handlers.
servers:
  - url: /api
paths:
...
  /admin/moderation/reports/{id}:
    patch:
      summary: PATCH /api/admin/moderation/reports/[id]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '429':
          description: Too Many Requests
        '500':
          description: Server Error
...
  /admin/users:
    get:
      summary: GET /api/admin/users
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '429':
          description: Too Many Requests
        '500':
          description: Server Error
    patch:
      summary: PATCH /api/admin/users
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '429':
          description: Too Many Requests
        '500':
          description: Server Error
    delete:
      summary: DELETE /api/admin/users
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '429':
          description: Too Many Requests
        '500':
          description: Server Error
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token